[Refactor/AdminPage] 옵션 카테고리 이름 컴포넌트 분리
