+++ src/components/AdminPage/AdminContainer.jsx
@@ -1,3 +1,4 @@
+import { useState } from "react";
 import { styled } from "styled-components";
 import Category from "./adminItems/Category";
 import OptionModal from "./adminOption/OptionModal";
@@ -11,6 +12,18 @@ const AdminContainerBox = styled.div`
 export default function AdminContainer({ numCategories }) {
   const categories = [];
 
+  /* 모달 관련 코드 일단 주석
+  const [isModalOpen, setIsModalOpen] = useState(false);
+
+  const openModal = () => {
+    setIsModalOpen(true);
+  };
+
+  const closeModal = () => {
+    setIsModalOpen(false);
+  };
+  */
+
   for (let i = 0; i < numCategories; i++) {
     categories.push(<Category key={i} />);
   }
@@ -18,6 +31,17 @@ export default function AdminContainer({ numCategories }) {
   return (
     <>
       <AdminContainerBox>
+        {/* <div
+          onClick={openModal}
+          style={{
+            padding: "10px",
+            background: "pink",
+            cursor: "pointer",
+          }}
+        >
+          옵션
+        </div>
+        <OptionModal isOpen={isModalOpen} onClose={closeModal} /> */}
         <OptionModal />
         {categories}
       </AdminContainerBox>
